ima: fix db size overflow and Kconfig issues
euleros inclusion
category: feature
bugzilla: https://gitee.com/openeuler/kernel/issues/I7QZ2M
CVE: NA

-----------------------------------------------------------------

This patch fixes some issues involved with IMA digest db size patches:
- use CONFIG_IMA_DIGEST_LIST to seperate db size check in ima_fs.c
- remove minimum digest db size limit in Kconfig
- change ima_digest_db_size data type from size_t to int, to avoid
  overflow
- change ima_digest_db_size in ima_add_digest_data_entry and
  ima_del_digest_data_entry
- add some debug messages

Fixes: 7c9d18bcaa ("ima: Add max size for IMA digest database")

+++ security/integrity/ima/ima.h
@@ -61,8 +61,8 @@ extern const char boot_aggregate_name[];
 extern int ima_digest_list_pcr;
 extern bool ima_plus_standard_pcr;
 extern int ima_digest_list_actions;
-extern size_t ima_digest_db_max_size __ro_after_init;
-extern size_t ima_digest_db_size;
+extern int ima_digest_db_max_size __ro_after_init;
+extern int ima_digest_db_size;
 #endif
 
 /* IMA event related data */
